+++ frontend/src/services/userServices.ts
@@ -347,6 +347,7 @@ export interface ModuleSettingsData {
   public_info: boolean;
   finance: boolean;
   medical_screening: boolean;
+  testing: boolean;
 }
 
 export interface OrganizationProfile {